requirements.txt
altgraph==0.17.4
blinker==1.8.2
bottle==0.12.25
cffi==1.16.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pefile==2023.2.7
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pythonnet==3.0.3
pywebview==5.1
pywin32-ctypes==0.2.2
setuptools==71.1.0
typing_extensions==4.12.2
Werkzeug==3.0.3
